Sync code with vanilla which handles particle swaps, Remove ability for sponge to "drag" elements
Originally pulled from tpt++: Add safeguard to try_move to ensure we don't write to out of bounds pmap 982fdff528dd
But the code in my mod was different and had extra checks. I looked back at history and found
3217151cd722
3ce18d29031b
1230fcd6acc5
b0402fef3169

The last one is notable because it says something about a crash with moving sponge. After looking at that code, it is complete janky nonsense, so I removed the feature to prevent it from causing more crashes.
Top commit also talked about stacking bugs on the edge of the screen. This is probably because of the unnecessary OutOfBounds checks which completely block swaps to the CELL border area and OOB areas. If these block a particle from swapping into OOB when it was supposed to do that and die, perhaps BHOL will build up.

Regarding removing the OutOfBounds checks, nx/ny are guaranteed to be in-bounds already. It's only rx/ry that may be out of bounds if a separate bug happens. So only one bounds check is needed