tpt++/LBPHacker (Update/Graphics lua changes) df7bbd3b8e30 254b4a642ab5
I added a copy of Simulation::elements called origElements, to be used for restoring update / graphics functions. The method used previously was calling the Init function again. This is not suitable for something that could be called thousands of times per frame.

I could replace the Init() call in sim.loadDefault() with origElements but will leave things as they are right now, I can't get rid of Init functions without major rework anyway.