Use Browse.json in save preview
This gives us save update timestamp, instead of a string, which we can use to print the update date in local time zone with the date format they select. This also gives us total comment count, so page count is finally accurate. There's still a limit of 200 loaded comments, might fix that in the next commit.