Add font hack, for compatibility with Lua scripts using vanilla tpt's icon range
Hack isn't as bad as I thought it would be, and is only applied to a few specific functions (gfx. ones, and tpt.drawstring, although not tpt.textwidth right now)
This is the last fix required to get "Alloy Brushes" working with no changes or special compatibility code needed in the script itself.

I didn't test the commit too much, but I also checked script manager, which now also works fine without compatibility for my mod, although the delete icon is offcentered somewhat. Offcentered icons are because many were deliberately adjusted somewhat in my mod, so that offsets wouldn't need to be applied on the c++ side to center it in certain buttons / locations. Not going to attempt to give compat for that.