tpt++/LBPHacker: Remove direct blocking prompt calls from Lua 4f31f85b6b29 59b79e805fc5 1b1ef99194d7 9605e0fcb90e 3ecc1c6c6fd3
Things of note:
I didn't *actually* remove anything in this commit. The legacy lua functions still exist for now, at least until after 98.0 releases. After that I'll probably remove them. After that, there will only be two blocking prompts left in new ui code. Old ui code is still full of them, but that's segmented into the legacy interface.cpp functions.

Using a legacy blocking prompt in the callback function a new prompt crashes the game. Not really sure why, or if other similar crashes exist.

TextPrompt was added to the new ui, and is used in a few places.

I implemented DeleteReason, which was mostly unused before. Buttons can also be set as Confirm buttons now. This just controls the reason given to the OnExit function. Nothing really checks this besides a few of the dialogs.

ConfirmPrompt's callback was changed a bit to mirror the rest of the prompt's new callbacks. Various code around the mod was changed to the new format.