Render / display mode refactors
Now tracked as a single bitfield instead of a list of bitfields. This is the same way vanilla does it now.
Now saved in powder.pref as a single bitfield as well. If the old format is there, it will load the list.
Fix air display modes not being mutually exclusive
Fix render / display modes not loading in tabs
Fix authors data possibly not loading sometimes (unconfirmed bug but I've seen it in practice)