tpt++: Implement Lua tools 00ec4e0754d1 554051736878
Some tool refactoring was also done to get things ready for it. More could have been done, but now tools have basic properties like "MenuSection" and "Color" they always should have had.

In the future (post-99.0), I don't plan on bringing new Lua APIs into my mod, exactly because of the difficulty in merging those changes. New Lua APIs often assume the codebase is in a good state, but this mod's codebase is so out of maintanence that everything is a challenge. It's not worth it to spend 2 days refactoring Tools to have basic functionality, or to fixup LuaSmartRef because some aspect of the new api requires it.